Remove `Pretty HsModule`
